refactor: apply codex window logic suggestions
